Actualizacion -_-

+++ PetCenter.Infrastucture.Domain.Main/BL_Planilla.cs
@@ -21,7 +21,7 @@ namespace PetCenter.Infrastucture.Domain.Main
             DA_SueldoMinimo DASueldoMinimo = new DA_SueldoMinimo();
             DA_Uit DAUit = new DA_Uit();
 
-            if (DAPlanilla.ExitePlanillaDeEsePeriodo(Fecha)) throw new Exception("PlanillaExiste");
+            if(DAPlanilla.ExitePlanillaDeEsePeriodo(Fecha)) throw new Exception("PlanillaExiste");
 
             Planilla planilla = new Planilla();
             try
@@ -35,7 +35,7 @@ namespace PetCenter.Infrastucture.Domain.Main
                 Calcular(planilla);
                 return DAPlanilla.ProcesarPlanilla(planilla);
             }
-            catch (Exception e)
+            catch(Exception e)
             {
                 EventLogger.EscribirLog(e.Message.ToString());
                 throw new Exception(e.Message.ToString());
@@ -49,7 +49,7 @@ namespace PetCenter.Infrastucture.Domain.Main
             {
                 return DAPlanilla.UltimaPlanilla(Fecha, PlanillaId);
             }
-            catch (Exception e)
+            catch(Exception e)
             {
                 EventLogger.EscribirLog(e.Message.ToString());
                 throw new Exception(e.Message.ToString());
@@ -64,7 +64,7 @@ namespace PetCenter.Infrastucture.Domain.Main
             List<Concepto> Conceptos = DaConceptos.ListarConceptos();
             List<Empleado> Empleados = DaEmpleados.ListarEmpleadosActivosDetallado();
 
-            foreach (var Empleado in Empleados)
+            foreach(var Empleado in Empleados)
             {
                 //Empleado.Faltas = DAFalta.ListarFaltasPorMesAndEmpleado(planilla.Fecha, Empleado.EmpleadoId);
                 decimal tot_I = 0.0m;
@@ -78,24 +78,24 @@ namespace PetCenter.Infrastucture.Domain.Main
                 //var conceptosfiltados = Conceptos;
                 PlanillaEmpleado planillaempleado = new PlanillaEmpleado();
 
-                foreach (var concepto in conceptosfiltados)
+                foreach(var concepto in conceptosfiltados)
                 {
                     PlanillaEmpleadoConcepto planillaempleadoconcepto = new PlanillaEmpleadoConcepto();
 
                     decimal calculo = getImporte(Empleado, concepto.calculo1, planilla.Fecha);
-                    if (!concepto.Operador1.Trim().Equals(""))
+                    if(!concepto.Operador1.Trim().Equals(""))
                     {
                         calculo = operar(Empleado, calculo, concepto.Operador1, concepto.calculo2);
-                        if (!concepto.Operador2.Trim().Equals(""))
+                        if(!concepto.Operador2.Trim().Equals(""))
                         {
                             calculo = operar(Empleado, calculo, concepto.Operador2, concepto.calculo3);
-                            if (!concepto.Operador3.Trim().Equals(""))
+                            if(!concepto.Operador3.Trim().Equals(""))
                             {
                                 calculo = operar(Empleado, calculo, concepto.Operador3, concepto.calculo4);
-                                if (!concepto.Operador4.Trim().Equals(""))
+                                if(!concepto.Operador4.Trim().Equals(""))
                                 {
                                     calculo = operar(Empleado, calculo, concepto.Operador4, concepto.calculo5);
-                                    if (!concepto.Operador5.Trim().Equals(""))
+                                    if(!concepto.Operador5.Trim().Equals(""))
                                     {
                                         calculo = operar(Empleado, calculo, concepto.Operador5, concepto.calculo6);
                                     }
@@ -105,48 +105,48 @@ namespace PetCenter.Infrastucture.Domain.Main
                     }
                     decimal valor = 0;
 
-                    if (concepto.Escala1 <= calculo)
+                    if(concepto.Escala1 <= calculo)
                     {
                         valor += concepto.Escala1 * concepto.Porcentaje1 / 100 + concepto.Importe1;
-                        if (concepto.Escala2 <= calculo)
+                        if(concepto.Escala2 <= calculo)
                         {
-                            valor += (concepto.Escala2 - concepto.Escala1) * concepto.Porcentaje2 / 100 + concepto.Importe2;
-                            if (concepto.Escala3 <= calculo)
+                            valor += (decimal)((concepto.Escala2 - concepto.Escala1) * concepto.Porcentaje2 / 100 + concepto.Importe2);
+                            if(concepto.Escala3 <= calculo)
                             {
-                                valor += (concepto.Escala3 - concepto.Escala2) * concepto.Porcentaje3 / 100 + concepto.Importe3;
-                                if (concepto.Escala4 <= calculo)
+                                valor += (decimal)((concepto.Escala3 - concepto.Escala2) * concepto.Porcentaje3 / 100 + concepto.Importe3);
+                                if(concepto.Escala4 <= calculo)
                                 {
-                                    valor += (concepto.Escala4 - concepto.Escala3) * concepto.Porcentaje4 / 100 + concepto.Importe4;
-                                    if (concepto.Escala5 <= calculo)
+                                    valor += (decimal)((concepto.Escala4 - concepto.Escala3) * concepto.Porcentaje4 / 100 + concepto.Importe4);
+                                    if(concepto.Escala5 <= calculo)
                                     {
-                                        valor += (concepto.Escala5 - concepto.Escala4) * concepto.Porcentaje5 / 100 + concepto.Importe5;
-                                        if (concepto.Escala6 <= calculo)
+                                        valor += (decimal)((concepto.Escala5 - concepto.Escala4) * concepto.Porcentaje5 / 100 + concepto.Importe5);
+                                        if(concepto.Escala6 <= calculo)
                                         {
-                                            valor += (concepto.Escala6 - concepto.Escala5) * concepto.Porcentaje6 / 100 + concepto.Importe6;
+                                            valor += (decimal)((concepto.Escala6 - concepto.Escala5) * concepto.Porcentaje6 / 100 + concepto.Importe6);
                                         }
                                         else
                                         {
-                                            valor += (calculo - concepto.Escala5) * concepto.Porcentaje6 / 100 + concepto.Importe6;
+                                            valor += (decimal)((calculo - concepto.Escala5) * concepto.Porcentaje6 / 100 + concepto.Importe6);
                                         }
                                     }
                                     else
                                     {
-                                        valor += (calculo - concepto.Escala4) * concepto.Porcentaje5 / 100 + concepto.Importe5;
+                                        valor += (decimal)((calculo - concepto.Escala4) * concepto.Porcentaje5 / 100 + concepto.Importe5);
                                     }
                                 }
                                 else
                                 {
-                                    valor += (calculo - concepto.Escala3) * concepto.Porcentaje4 / 100 + concepto.Importe4;
+                                    valor += (decimal)((calculo - concepto.Escala3) * concepto.Porcentaje4 / 100 + concepto.Importe4);
                                 }
                             }
                             else
                             {
-                                valor += (calculo - concepto.Escala2) * concepto.Porcentaje3 / 100 + concepto.Importe3;
+                                valor += (decimal)((calculo - concepto.Escala2) * concepto.Porcentaje3 / 100 + concepto.Importe3);
                             }
                         }
                         else
                         {
-                            valor += (calculo - concepto.Escala1) * concepto.Porcentaje2 / 100 + concepto.Importe2;
+                            valor += (decimal)((calculo - concepto.Escala1) * concepto.Porcentaje2 / 100 + concepto.Importe2);
                         }
                     }
                     else
@@ -154,12 +154,12 @@ namespace PetCenter.Infrastucture.Domain.Main
                         valor += calculo * concepto.Porcentaje1 / 100 + concepto.Importe1;
                     }
 
-                    if (concepto.calculo1.Equals("12SUBA"))
+                    if(concepto.calculo1.Equals("12SUBA"))
                     {
                         valor /= 12;
                     }
 
-                    switch (concepto.Tipo)
+                    switch(concepto.Tipo)
                     {
                         case 0:
                             tot_I += valor;
@@ -174,7 +174,7 @@ namespace PetCenter.Infrastucture.Domain.Main
                     planillaempleadoconcepto.Importe = valor;
                     planillaempleadoconcepto.ConceptoId = concepto.ConceptoId;
 
-                    if (planillaempleado.PlanillaEmpleadoConceptoes == null) planillaempleado.PlanillaEmpleadoConceptoes = new List<PlanillaEmpleadoConcepto>();
+                    if(planillaempleado.PlanillaEmpleadoConceptoes == null) planillaempleado.PlanillaEmpleadoConceptoes = new List<PlanillaEmpleadoConcepto>();
                     planillaempleado.PlanillaEmpleadoConceptoes.Add(planillaempleadoconcepto);
                 }
                 planillaempleado.TotalAporte = tot_A;
@@ -185,62 +185,62 @@ namespace PetCenter.Infrastucture.Domain.Main
                 planillaempleado.AfpId = 1;
                 planillaempleado.OnpId = 1;
 
-                if (planilla.PlanillaEmpleadoes == null) planilla.PlanillaEmpleadoes = new List<PlanillaEmpleado>();
+                if(planilla.PlanillaEmpleadoes == null) planilla.PlanillaEmpleadoes = new List<PlanillaEmpleado>();
                 planilla.PlanillaEmpleadoes.Add(planillaempleado);
             }
         }
         private static decimal getImporte(Empleado empleado, String campo, DateTime? Fecha = null)
         {
             decimal importe = 0;
-            if (campo.Equals("TOTFAL"))
+            if(campo.Equals("TOTFAL"))
             {
                 importe = getTotalFaltas(empleado);
             }
-            else if (campo.Equals("SUEBAS"))
+            else if(campo.Equals("SUEBAS"))
             {
                 importe = getSueldoBase(empleado);
             }
-            else if (campo.Equals("TOTTAR"))
+            else if(campo.Equals("TOTTAR"))
             {
                 importe = getTotalTardanzas(empleado, (DateTime)Fecha);
             }
-            else if (campo.Equals("SUEMIN"))
+            else if(campo.Equals("SUEMIN"))
             {
                 importe = getSueldoMinimo();
             }
-            else if (campo.Equals("APOAFP"))
+            else if(campo.Equals("APOAFP"))
             {
                 importe = getAporteAfp(empleado);
             }
-            else if (campo.Equals("SEGAFP"))
+            else if(campo.Equals("SEGAFP"))
             {
                 importe = getSeguroAfp(empleado);
             }
-            else if (campo.Equals("COMAFP"))
+            else if(campo.Equals("COMAFP"))
             {
                 importe = getComisionAfp(empleado);
             }
-            else if (campo.Equals("APOONP"))
+            else if(campo.Equals("APOONP"))
             {
                 importe = getAporteOnp();
             }
-            else if (campo.Equals("APOESA"))
+            else if(campo.Equals("APOESA"))
             {
                 importe = getAporteEssalud();
             }
-            else if (campo.Equals("SEGVIL"))
+            else if(campo.Equals("SEGVIL"))
             {
                 importe = getSeguroVidaLey();
             }
-            else if (campo.Equals("12SUBA"))
+            else if(campo.Equals("12SUBA"))
             {
                 importe = getImpuestoRenta(empleado);
             }
-            else if (campo.Equals("GRATIF"))
+            else if(campo.Equals("GRATIF"))
             {
                 importe = getGratificacion(empleado);
             }
-            else if (isNumeric(campo))
+            else if(isNumeric(campo))
             {
                 importe = Convert.ToDecimal(campo);
             }
@@ -255,7 +255,7 @@ namespace PetCenter.Infrastucture.Domain.Main
             {
                 return DAPlanilla.findPlanilla(Fecha);
             }
-            catch (Exception e)
+            catch(Exception e)
             {
                 EventLogger.EscribirLog(e.Message.ToString());
                 throw new Exception(e.Message.ToString());
@@ -269,7 +269,7 @@ namespace PetCenter.Infrastucture.Domain.Main
             {
                 return DAPlanilla.delPlanilla(planilla);
             }
-            catch (Exception e)
+            catch(Exception e)
             {
                 EventLogger.EscribirLog(e.Message.ToString());
                 throw new Exception(e.Message.ToString());
@@ -278,30 +278,30 @@ namespace PetCenter.Infrastucture.Domain.Main
 
         private static decimal getTotalFaltas(Empleado empleado)
         {
-            if (empleado.Faltas == null || empleado.Faltas.Count <= 0) return 0.0m;
+            if(empleado.Faltas == null || empleado.Faltas.Count <= 0) return 0.0m;
 
             return empleado.Faltas.Count();
         }
 
         private static decimal getSueldoBase(Empleado empleado)
         {
-            if (empleado.EmpleadoSueldoBases == null || empleado.EmpleadoSueldoBases.Count <= 0) return 0.0m;
+            if(empleado.EmpleadoSueldoBases == null || empleado.EmpleadoSueldoBases.Count <= 0) return 0.0m;
             var item = empleado.EmpleadoSueldoBases.Where(s => s.Estado).FirstOrDefault();
-            if (item == null) return 0.0m;
+            if(item == null) return 0.0m;
             return item.SueldoBase;
         }
 
         private static decimal getTotalTardanzas(Empleado empleado, DateTime Fecha)
         {
-            if (empleado.Tardanzas == null || empleado.Tardanzas.Count <= 0) return 0.0m;
+            if(empleado.Tardanzas == null || empleado.Tardanzas.Count <= 0) return 0.0m;
             var items = empleado.Tardanzas.Where(s => s.FechaTardanza.Month.Equals(Fecha.Month) && s.FechaTardanza.Year.Equals(Fecha.Year));
-            if (items == null) return 0.0m;
+            if(items == null) return 0.0m;
             return Convert.ToDecimal(items.Sum(s => s.Minutos));
         }
 
         public static Boolean isNumeric(String str)
         {
-            if (str == null)
+            if(str == null)
             {
                 return false;
             }
@@ -311,7 +311,7 @@ namespace PetCenter.Infrastucture.Domain.Main
 
         private static decimal operar(Empleado empleado, decimal importe, String operador, String calculo)
         {
-            switch (operador)
+            switch(operador)
             {
                 case "+":
                     importe += getImporte(empleado, calculo);
@@ -333,37 +333,37 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             DA_SueldoMinimo DASueldoMinimo = new DA_SueldoMinimo();
             var item = DASueldoMinimo.SueldoMinimoActivo();
-            if (item == null) return 0.0m;
+            if(item == null) return 0.0m;
             return item.Valor;
         }
 
         private static decimal getAporteAfp(Empleado empleado)
         {
-            if (empleado.EmpleadoAfps == null) return 0.0m;
+            if(empleado.EmpleadoAfps == null) return 0.0m;
             var items = empleado.EmpleadoAfps.Where(s => s.Estado).FirstOrDefault();
-            if (items == null) return 0.0m;
+            if(items == null) return 0.0m;
             var afp = items.Afp;
-            if (afp == null) return 0.0m;
+            if(afp == null) return 0.0m;
             return afp.Aporte;
         }
 
         private static decimal getSeguroAfp(Empleado empleado)
         {
-            if (empleado.EmpleadoAfps == null) return 0.0m;
+            if(empleado.EmpleadoAfps == null) return 0.0m;
             var items = empleado.EmpleadoAfps.Where(s => s.Estado).FirstOrDefault();
-            if (items == null) return 0.0m;
+            if(items == null) return 0.0m;
             var afp = items.Afp;
-            if (afp == null) return 0.0m;
+            if(afp == null) return 0.0m;
             return afp.Seguro;
         }
 
         private static decimal getComisionAfp(Empleado empleado)
         {
-            if (empleado.EmpleadoAfps == null) return 0.0m;
+            if(empleado.EmpleadoAfps == null) return 0.0m;
             var items = empleado.EmpleadoAfps.Where(s => s.Estado).FirstOrDefault();
-            if (items == null) return 0.0m;
+            if(items == null) return 0.0m;
             var afp = items.Afp;
-            if (afp == null) return 0.0m;
+            if(afp == null) return 0.0m;
             return afp.Comision;
         }
 
@@ -371,7 +371,7 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             DA_Onp DAOnp = new DA_Onp();
             var item = DAOnp.OnpActivo();
-            if (item == null) return 0.0m;
+            if(item == null) return 0.0m;
             return item.Aporte;
         }
 
@@ -379,7 +379,7 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             DA_EsSalud DA_EsSalud = new DA_EsSalud();
             var essalud = DA_EsSalud.EsSaludActivo();
-            if (essalud == null) return 0.0m;
+            if(essalud == null) return 0.0m;
             return essalud.Valor;
 
         }
@@ -388,7 +388,7 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             DA_SeguroVidaLey DASeguroVidaLey = new DA_SeguroVidaLey();
             var segurovidaley = DASeguroVidaLey.SeguroVidaLeyActiva();
-            if (segurovidaley == null) return 0.0m;
+            if(segurovidaley == null) return 0.0m;
             return segurovidaley.Valor;
         }
 
@@ -396,7 +396,7 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             decimal remuneracion = 12 * getSueldoBase(empleado) + 2 * getGratificacion(empleado);
             remuneracion = remuneracion - 7 * getUIT();
-            if (remuneracion < 0) remuneracion = 0;
+            if(remuneracion < 0) remuneracion = 0;
             return remuneracion;
         }
 
@@ -409,7 +409,7 @@ namespace PetCenter.Infrastucture.Domain.Main
         {
             DA_Uit DAUit = new DA_Uit();
             var item = DAUit.UitActiva();
-            if (item == null) return 0;
+            if(item == null) return 0;
             return Convert.ToInt32(item.Valor);
         }
     }
